chore: convert app-definition
